app/user/permissions.py
from rest_framework import permissions


class OwnerUserPerms(permissions.IsAuthenticated):
    def has_object_permission(self, request, view, obj):
        """
        Return `True` if permission is granted, `False` otherwise.
        """
        return request.user == obj
