app/requirements.txt
# Main
Django==4.2.4
djangorestframework==3.14.0
django-oauth-toolkit==2.3.0

# SQL Server
psycopg2==2.9.6
mysqlclient==2.1.1


Pillow==9.5.0
django-environ==0.10.0
whitenoise==6.4.0


# Celery
celery==5.2.7
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.0.0

# Network and social
facebook-sdk==3.1.0
gevent==22.10.2
google-api-core==2.11.0
google-api-python-client==2.88.0


# Tool
daphne==4.0.0 # ASGI server
twisted[tls,http2]
polars==0.18.0
requests==2.31.0
django-money==3.1.0
PyJWT==2.7.0
XlsxWriter==3.1.2
drf-yasg==1.21.5
django-debug-toolbar==4.1.0
django-extensions==3.2.3
ipython==8.14.0
xlsx2csv==0.8.1

# Redis
redis==4.5.5
django-redis==5.2.0
channels-redis==4.1.0
django-sse==0.4.1
sse==1.2.0